Examples: 09 localizer: fix use of boolean attributes
Fix the use of boolean attributes, which are not allowed to have values
assigned to them (i.e., "false" or "true" are incorrect).  See for
example
https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes#Boolean_Attributes,
which states:

"To be clear, the values "true" and "false" are not allowed on boolean
attributes. To represent a false value, the attribute has to be omitted
altogether. This restriction clears up some common misunderstandings:
With checked="false" for example, the element’s checked attribute would
be interpreted as true because the attribute is present."